server/requirements.txt
fastapi
uvicorn
motor
certifi
python-dateutil